backend/requirements.txt
flask==2.3.3
redis==4.6.0
mysql-connector-python==8.1.0
